+++ kornia/contrib/edge_detection.py
@@ -1,8 +1,8 @@
 from typing import List
 
 from kornia.core import Module, Tensor
+from kornia.core.logger import KORNIA_CHECK_SHAPE
 from kornia.filters.dexined import DexiNed
-from kornia.testing import KORNIA_CHECK_SHAPE
 
 
 class EdgeDetector(Module):
