+++ kornia/enhance/shift_rgb.py
@@ -1,6 +1,6 @@
 import torch
 
-from kornia.testing import KORNIA_CHECK_IS_COLOR, KORNIA_CHECK_IS_TENSOR
+from kornia.core.logger import KORNIA_CHECK_IS_COLOR, KORNIA_CHECK_IS_TENSOR
 
 
 def shift_rgb(image: torch.Tensor, r_shift: torch.Tensor, g_shift: torch.Tensor, b_shift: torch.Tensor) -> torch.Tensor:
