+++ kornia/feature/integrated.py
@@ -5,9 +5,9 @@ import torch
 
 from kornia.color import rgb_to_grayscale
 from kornia.core import Device, Module, Tensor, concatenate
+from kornia.core.logger import KORNIA_CHECK_LAF
 from kornia.geometry.subpix import ConvQuadInterp3d
 from kornia.geometry.transform import ScalePyramid
-from kornia.testing import KORNIA_CHECK_LAF
 
 from .affine_shape import LAFAffNetShapeEstimator
 from .hardnet import HardNet
