+++ kornia/feature/responses.py
@@ -3,8 +3,8 @@ from typing import Optional, Union
 import torch
 
 from kornia.core import Module, Tensor, tensor
+from kornia.core.logger import KORNIA_CHECK_SHAPE
 from kornia.filters import gaussian_blur2d, spatial_gradient
-from kornia.testing import KORNIA_CHECK_SHAPE
 
 
 def harris_response(
