+++ kornia/feature/sold2/sold2_detector.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, Optional, Tuple
 import torch
 
 from kornia.core import Module, Tensor, concatenate, stack, tensor, where, zeros
+from kornia.core.logger import KORNIA_CHECK_SHAPE
 from kornia.geometry.bbox import nms
-from kornia.testing import KORNIA_CHECK_SHAPE
 from kornia.utils import map_location_to_cpu, torch_meshgrid
 
 from .backbones import SOLD2Net
