+++ kornia/feature/tfeat.py
@@ -3,7 +3,7 @@ from typing import Dict
 import torch
 import torch.nn as nn
 
-from kornia.testing import KORNIA_CHECK_SHAPE
+from kornia.core.logger import KORNIA_CHECK_SHAPE
 from kornia.utils.helpers import map_location_to_cpu
 
 urls: Dict[str, str] = {}
