+++ kornia/geometry/epipolar/_metrics.py
@@ -2,9 +2,9 @@
 
 from torch import Tensor
 
+from kornia.core.logger import KORNIA_CHECK_IS_TENSOR
 from kornia.geometry.conversions import convert_points_to_homogeneous
 from kornia.geometry.linalg import point_line_distance
-from kornia.testing import KORNIA_CHECK_IS_TENSOR
 
 
 def sampson_epipolar_distance(
