+++ kornia/geometry/epipolar/fundamental.py
@@ -5,9 +5,9 @@ from typing import Optional, Tuple
 import torch
 
 from kornia.core import Tensor
+from kornia.core.logger import KORNIA_CHECK_SHAPE
 from kornia.geometry.conversions import convert_points_from_homogeneous, convert_points_to_homogeneous
 from kornia.geometry.linalg import transform_points
-from kornia.testing import KORNIA_CHECK_SHAPE
 from kornia.utils.helpers import _torch_svd_cast
