+++ kornia/losses/cauchy.py
@@ -1,7 +1,7 @@
 from torch import Tensor
 
 from kornia.core import Module
-from kornia.testing import KORNIA_CHECK, KORNIA_CHECK_IS_TENSOR, KORNIA_CHECK_SAME_DEVICE, KORNIA_CHECK_SAME_SHAPE
+from kornia.core.logger import KORNIA_CHECK, KORNIA_CHECK_IS_TENSOR, KORNIA_CHECK_SAME_DEVICE, KORNIA_CHECK_SAME_SHAPE
 
 
 def cauchy_loss(img1: Tensor, img2: Tensor, reduction: str = "none") -> Tensor:
