+++ kornia/losses/lovasz_softmax.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from kornia.testing import KORNIA_CHECK_SHAPE
+from kornia.core.logger import KORNIA_CHECK_SHAPE
 
 # based on:
 # https://github.com/bermanmaxim/LovaszSoftmax
